Chatbot-with-OpenAI/requirements.txt
openai 
chainlit
python-dotenv